requirements.txt
# Core data science and ML libraries
numpy>=1.21.0,<2.0.0
pandas>=1.5.0,<3.0.0
scikit-learn>=1.1.0,<2.0.0
joblib>=1.2.0,<2.0.0

# Plotting and visualization
matplotlib>=3.5.0,<4.0.0
seaborn>=0.11.0,<1.0.0
plotly>=5.0.0,<6.0.0

# Machine learning models
xgboost>=1.6.0,<2.0.0

# Deep learning (Apple Silicon optimized for M4)
tensorflow-macos>=2.10.0,<3.0.0
tensorflow-metal>=1.0.0,<2.0.0

# Explainability and interpretability
shap>=0.40.0,<1.0.0
lime>=0.2.0,<1.0.0

# Web application framework
streamlit>=1.25.0,<2.0.0

# Data formats and processing
pyarrow>=10.0.0,<15.0.0
psutil>=5.8.0,<6.0.0

# Real-time processing and async
aiofiles>=22.0.0,<25.0.0

# Database connectivity
psycopg2-binary>=2.9.0,<3.0.0
redis>=4.5.0,<5.0.0
sqlalchemy>=1.4.0,<3.0.0

# API framework
fastapi>=0.100.0,<1.0.0
uvicorn>=0.20.0,<1.0.0
pydantic>=1.10.0,<3.0.0

# HTTP requests and communication
requests>=2.28.0,<3.0.0
httpx>=0.24.0,<1.0.0

# Configuration and environment
python-dotenv>=0.19.0,<2.0.0
pyyaml>=6.0.0,<7.0.0

# Logging and monitoring
structlog>=22.0.0,<24.0.0

# Security and encryption
cryptography>=38.0.0,<43.0.0

# Testing and quality assurance
pytest>=7.0.0,<8.0.0
pytest-asyncio>=0.20.0,<1.0.0
pytest-cov>=4.0.0,<5.0.0
black>=22.0.0,<25.0.0
flake8>=5.0.0,<7.0.0

# Development and deployment
docker>=6.0.0,<7.0.0

# Medical and clinical data processing
pydicom>=2.3.0,<3.0.0

# Time series analysis
statsmodels>=0.13.0,<1.0.0

# Additional utilities
tqdm>=4.60.0,<5.0.0
click>=8.0.0,<9.0.0
rich>=12.0.0,<14.0.0

# macOS M4 specific optimizations
# Note: Some packages may need to be installed with specific flags for M4
# For example: pip install tensorflow-macos --no-deps
# Then: pip install tensorflow-metal --no-deps
